punnet_aacharya/requirements.txt
python-telegram-bot==20.7
pyswisseph==2.10.3.2
pytz==2024.1
geopy==2.4.1
timezonefinder==6.4.0
sqlalchemy==2.0.25
chromadb==0.4.22
pandas==2.1.4
numpy==1.26.3
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
scipy==1.11.4
google-generativeai==0.8.5 